requirements.txt
streamlit~=1.27.0
pandas~=2.1.0
tensorflow~=2.13.0
numpy~=1.24.3
mysql-connector-python~=8.1.0
requests~=2.31.0
pillow~=9.4.0
langchain~=0.0.302
beautifulsoup4~=4.12.2
matplotlib
google-search-results


